+++ src/app/inventory/store/d1-item-factory.ts
@@ -362,7 +362,6 @@ function makeItem(
     energy: null,
     powerCap: null,
     pursuit: null,
-    bungieIndex: 0,
   };
 
   // *able