+++ src/app/progress/milestone-items.ts
@@ -291,7 +291,6 @@ function makeFakePursuitItem(
     trackable: false,
     energy: null,
     powerCap: null,
-    bungieIndex: 0,
   };
 }