+++ src/app/vendors/d2-vendors.ts
@@ -2,7 +2,7 @@ import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { ItemCreationContext } from 'app/inventory/store/d2-item-factory';
 import { VENDORS, silverItemHash } from 'app/search/d2-known-values';
 import { ItemFilter } from 'app/search/filter-types';
-import { compareBy } from 'app/utils/comparators';
+import { chainComparator, compareBy } from 'app/utils/comparators';
 import { filterMap } from 'app/utils/util';
 import {
   DestinyCollectibleState,
@@ -94,6 +94,16 @@ export function toVendor(
   }
 
   const vendorItems = getVendorItems(context, vendorDef, characterId, sales);
+  vendorItems.sort(
+    chainComparator(
+      compareBy(
+        (item) =>
+          item.originalCategoryIndex !== undefined &&
+          vendorDef.originalCategories[item.originalCategoryIndex]?.sortValue
+      ),
+      compareBy((item) => item.vendorItemIndex)
+    )
+  );
 
   const destinationDef =
     typeof vendor?.vendorLocationIndex === 'number' && vendor.vendorLocationIndex >= 0